requirements.txt
pandas==2.3.2
numpy==2.1.3
matplotlib==3.9.2
python-dotenv==1.0.1
openai==1.52.0
scipy==1.14.1
alpha-vantage==3.0.0
yfinance==0.2.43
anthropic>=0.18.0
reportlab==4.4.9
pillow==11.1.0